Moved web implementation files to subfolders
